Move additional functions to tblex
Also avoids circular import between tblex and treebankfunctions.
