+++ tblex.py
@@ -1,10 +1,16 @@
 """
-The module *tblex* contains functiond that require function from the lexicon module and from the treebankfunctions module
+The module *tblex* contains functiond that require function
+from the lexicon module and from the treebankfunctions module
 """
 
 from sastatypes import SynTree
 import lexicon as lex
-from treebankfunctions import getattval, iscompound, isdiminutive
+from treebankfunctions import (all_lower_consonantsnode, getattval,
+                               is_duplicate_spec_noun, iscompound,
+                               isdiminutive, isnumber, issubstantivised_verb,
+                               sasta_long, sasta_pseudonym,
+                               short_nucl_n, spec_noun)
+
 
 def recognised_wordnodepos(node: SynTree, pos: str) -> bool:
     '''
@@ -34,8 +40,9 @@ def recognised_wordnodepos(node: SynTree, pos: str) -> bool:
     word = getattval(node, 'word')
     lcword = word.lower()
     result = lex.informlexiconpos(word, pos) or lex.informlexiconpos(lcword, pos) or \
-             iscompound(node) or isdiminutive(node) or lex.isa_namepart_uc(word)
+        iscompound(node) or isdiminutive(node) or lex.isa_namepart_uc(word)
     return result
+
 
 def recognised_wordnode(node: SynTree) -> bool:
     '''
@@ -64,8 +71,11 @@ def recognised_wordnode(node: SynTree) -> bool:
 
     word = getattval(node, 'word')
     lcword = word.lower()
-    result = lex.informlexicon(word) or lex.informlexicon(lcword) or iscompound(node) or isdiminutive(
-        node) or lex.isa_namepart(word)
+    result = lex.informlexicon(word) \
+        or lex.informlexicon(lcword) \
+        or iscompound(node) \
+        or isdiminutive(node) \
+        or lex.isa_namepart(word)
     return result
 
 
@@ -90,3 +100,91 @@ def recognised_lemmanodepos(node: SynTree, pos: str) -> bool:
     result = lex.informlexiconpos(lemma, pos)
     return result
 
+
+def asta_recognised_lexnode(node: SynTree) -> bool:
+    '''
+    The function *asta_recognised_lexnode* determines whether *node* should count as a
+    lexical verb in the ASTA method.
+
+    This is the case if *pt* equals *ww* and the node is not a substantivised verb as
+    determined by the function *issubstantivised_verb*:
+
+    .. autofunction:: treebankfunctions::issubstantivised_verb
+
+    '''
+    if issubstantivised_verb(node):
+        result = False
+    else:
+        result = getattval(node, 'pt') == 'ww'
+    return result
+
+
+def asta_recognised_nounnode(node: SynTree) -> bool:
+    '''
+    The function *asta_recognised_nounnode* determines whether *node* should count as a
+    noun in the ASTA method.
+
+    This is the case if
+
+    * either the node meets the conditions of *sasta_pseudonym*
+
+       .. autofunction:: treebankfunctions::sasta_pseudonym
+
+    * or the node meets the conditions of *spec_noun*
+
+       .. autofunction:: treebankfunctions::spec_noun
+
+    * or the node meets the conditions of *is_duplicate_spec_noun*
+
+       .. autofunction:: treebankfunctions::is_duplicate_spec_noun
+
+    * or the node meets the conditions of *sasta_long*
+
+       .. autofunction:: treebankfunctions::sasta_long
+
+    * or the node meets the conditions of *recognised_wordnodepos*
+
+       .. autofunction:: treebankfunctions::recognised_wordnodepos
+
+    * or the node meets the conditions of *recognised_lemmanodepos(node, pos)*
+
+       .. autofunction:: treebankfunctions::recognised_lemmanodepos(node, pos)
+
+    However, the node should:
+
+    * neither consist of lower case consonants only, as determined by *all_lower_consonantsnode*:
+
+       .. autofunction:: treebankfunctions::all_lower_consonantsnode
+
+    * nor satisfy the conditions of *short_nucl_n*:
+
+       .. autofunction:: treebankfunctions::short_nucl_n
+
+    '''
+
+    if issubstantivised_verb(node):
+        pos = 'ww'
+    else:
+        pos = 'n'
+    result = sasta_pseudonym(node)
+    result = result or spec_noun(node)
+    result = result or is_duplicate_spec_noun(node)
+    result = result or sasta_long(node)
+    result = result or recognised_wordnodepos(node, pos)
+    result = result or recognised_lemmanodepos(node, pos)
+    result = result and not (all_lower_consonantsnode(node))
+    result = result and not (short_nucl_n(node))
+    return result
+
+
+def asta_recognised_wordnode(node: SynTree) -> bool:
+    result = sasta_pseudonym(node)
+    result = result or spec_noun(node)
+    result = result or is_duplicate_spec_noun(node)
+    result = result or sasta_long(node)
+    result = result or recognised_wordnode(node)
+    result = result or recognised_lemmanode(node)
+    result = result or isnumber(node)
+    result = result and not (all_lower_consonantsnode(node))
+    result = result and not (short_nucl_n(node))
+    return result
